requirements.txt
deepmerge>=1.1.0
dotmap>=1.3.30
jsonschema>=4.17.3
matplotlib>=3.7.1
numpy>=1.25.0
pandas>=2.0.2
rich>=13.4.2
screeninfo>=0.8.1
fastapi>=0.110.0
uvicorn>=0.29.0
jinja2>=3.1.4
python-multipart>=0.0.9
itsdangerous>=2.2.0
reportlab>=4.2.0
pillow>=10.3.0
opencv-python-headless>=4.9.0.80
